refactor: move some functions to utils/items
